+++ src/resources/views/admin/element/form/input-multi-lang.blade.php
@@ -8,32 +8,34 @@
                 <a class="nav-link dropdown-toggle show-item-change-language-{{$name}}" data-toggle="dropdown" href="#"
                    role="button"
                    aria-haspopup="true" aria-expanded="false" style="text-transform: uppercase">
-                    {{ config('app.locale') }}
+                    {{ $language_content['vi'] }}
                 </a>
-
-                <div class="dropdown-menu" style="margin: 0">
-                    @foreach(config('app.language_text') as $lang => $textLanguage)
-                        <a data-toggle="collapse"
-                           data-code="{{$lang}}"
-                           data-parent="#myGroup-{{$name}}"
-                           href="#input-lang-{{$name}}-{{$lang}}"
-                           class="dropdown-item dropdown-item-change-language-{{$name}}">
-                            {{ $textLanguage }}
-                        </a>
-                    @endforeach
-                </div>
+                @if(count($language_content) > 1)
+                    <div class="dropdown-menu" style="margin: 0">
+                        @foreach($language_content as $lang => $textLanguage)
+                            <a data-toggle="collapse"
+                               data-code="{{$lang}}"
+                               data-text="{{$textLanguage}}"
+                               data-parent="#myGroup-{{$name}}"
+                               href="#input-lang-{{$name}}-{{$lang}}"
+                               class="dropdown-item dropdown-item-change-language-{{$name}}">
+                                {{ $textLanguage }}
+                            </a>
+                        @endforeach
+                    </div>
+                @endif
             </li>
         </ul>
     </div>
     <div class="accordion-group">
-        @foreach(config('app.language_text') as $lang => $textLanguage)
+        @foreach($language_content as $lang => $textLanguage)
             <div id="input-lang-{{$name}}-{{$lang}}"
-                 class="collapse collapse-item-change-language-{{$name}} @if($lang == config('app.locale')) show @endif">
+                 class="collapse collapse-item-change-language-{{$name}} @if($lang == 'vi') show @endif">
                 <div class="input-group">
                     <input type="{{ $type ?? 'text' }}"
                            name="{{$lang}}[{{$name}}]"
                            id="{{$name}}"
-                           placeholder="[{{$text}}: {{ $textLanguage }}]"
+                           placeholder="{{$text}}: {{ $textLanguage }}"
                            value="{{ old($name, $value->translate($lang)->$name ?? '') }}"
                            class="form-control {{ $class ?? '' }}"
                            autocomplete="off"/>
@@ -52,8 +54,8 @@
 <script type="text/javascript">
     jQuery('.dropdown-item-change-language-{{$name}}').click(function (e) {
         jQuery('.collapse-item-change-language-{{$name}}').collapse('hide');
-        let code = $(this).attr('data-code');
-        jQuery('.show-item-change-language-{{$name}}').text(code);
+        let text = $(this).attr('data-text');
+        jQuery('.show-item-change-language-{{$name}}').text(text);
     });
 </script>