+++ src/resources/views/admin/element/form/textarea-multi-lang.blade.php
@@ -5,33 +5,37 @@
     <div class="input-language-change">
         <ul class="nav nav-pills">
             <li class="nav-item dropdown">
-                <a class="nav-link dropdown-toggle show-item-change-language-textarea-{{$name}}" data-toggle="dropdown" href="#"
+                <a class="nav-link dropdown-toggle show-item-change-language-textarea-{{$name}}" data-toggle="dropdown"
+                   href="#"
                    role="button"
                    aria-haspopup="true" aria-expanded="false" style="text-transform: uppercase">
-                    {{ config('app.locale') }}
+                    {{ $language_content['vi'] }}
                 </a>
 
-                <div class="dropdown-menu" style="margin: 0">
-                    @foreach(config('app.language_text') as $lang => $textLanguage)
-                        <a data-toggle="collapse"
-                           data-code="{{$lang}}"
-                           data-parent="#myGroup-{{$name}}"
-                           href="#input-lang-{{$name}}-{{$lang}}"
-                           class="dropdown-item dropdown-item-change-language-textarea-{{$name}}">
-                            {{ $textLanguage }}
-                        </a>
-                    @endforeach
-                </div>
+                @if(count($language_content) > 1)
+                    <div class="dropdown-menu" style="margin: 0">
+                        @foreach($language_content as $lang => $textLanguage)
+                            <a data-toggle="collapse"
+                               data-code="{{$lang}}"
+                               data-text="{{$textLanguage}}"
+                               data-parent="#myGroup-{{$name}}"
+                               href="#input-lang-{{$name}}-{{$lang}}"
+                               class="dropdown-item dropdown-item-change-language-textarea-{{$name}}">
+                                {{ $textLanguage }}
+                            </a>
+                        @endforeach
+                    </div>
+                @endif
             </li>
         </ul>
     </div>
     <div class="accordion-group">
-        @foreach(config('app.language_text') as $lang => $textLanguage)
+        @foreach($language_content as $lang => $textLanguage)
             <div id="input-lang-{{$name}}-{{$lang}}"
-                 class="collapse collapse-item-change-language-textarea-{{$name}} @if($lang == config('app.locale')) show @endif">
+                 class="collapse collapse-item-change-language-textarea-{{$name}} @if($lang == 'vi') show @endif">
                 <div class="input-group">
                     <textarea name="{{$lang}}[{{$name}}]"
-                              placeholder="[{{$text}}: {{ $textLanguage }}]"
+                              placeholder="{{$text}}: {{ $textLanguage }}"
                               id="input-lang-{{$name}}-{{$lang}}"
                               rows="{{ $rows ?? 5 }}"
                               class="form-control {{ $class ?? '' }}">{{ old($name, $value->translate($lang)->$name ?? '') }}</textarea>
@@ -50,7 +54,7 @@
 <script type="text/javascript">
     jQuery('.dropdown-item-change-language-textarea-{{$name}}').click(function (e) {
         jQuery('.collapse-item-change-language-textarea-{{$name}}').collapse('hide');
-        let code = $(this).attr('data-code');
-        jQuery('.show-item-change-language-textarea-{{$name}}').text(code);
+        let text = $(this).attr('data-text');
+        jQuery('.show-item-change-language-textarea-{{$name}}').text(text);
     });
 </script>