+++ CHANGES.md
@@ -10,3 +10,5 @@ Laravel Wordpress Cms Changelog
 		+ GOOGLE_RECAPTCHA_KEY=
 		+ GOOGLE_RECAPTCHA_SECRET=
 	- how to on/off google recaptcha GOOGLE_RECAPTCHA_ENABLED=true/false
+
+* 2021-03-30: add manager multi language site